requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
openai>=1.0.0
httpx>=0.27.0
hypothesis>=6.100.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pyyaml>=6.0.0
starlette>=0.36.0
